Swapped synopsis & description in .cabal file
It didn't really make sense showing the libgit2 version number in the
synopsis. It seems like it is better suited to the description field so
that you only see it when you look at the hlibgit2 page directly.

I also changed 'Generated bindings' to 'Low-level bindings' in
accordance with many of the other bindings on Hackage. Perhaps we should
rename the library to bindings-libgit2? This seems to be a common naming
scheme for bindings on Hackage.