More refactoring on the S3 code
This is being done because it had become something of a mess, and was
hard to reason about the behavior.  As part of making it more resilient
to database failures (i.e., to allow the S3 data to "standalone", if
necessary), it was necessary to clean up the code to ensure that every
possible contingency is coded for.